Корректировки в соответствии с замечаниями ревьюера
