+++ MovieQuiz/Helpers/UIColor+Extensions.swift
@@ -1,10 +1,10 @@
 import UIKit
 
 extension UIColor {
-     static var ypGreen: UIColor { UIColor(named: "YP Green") ?? UIColor.green }
-     static var ypRed: UIColor { UIColor(named: "YP Red") ?? UIColor.red }
-     static var ypBlack: UIColor { UIColor(named: "YP Black") ?? UIColor.black }
-     static var ypBackground: UIColor { UIColor(named: "YP Background") ?? UIColor.darkGray }
-     static var ypGray: UIColor { UIColor(named: "YP Gray") ?? UIColor.gray }
-     static var ypWhite: UIColor { UIColor(named: "YP White") ?? UIColor.white}
+    static var ypGreen: UIColor { UIColor(named: "YP Green") ?? UIColor.green }
+    static var ypRed: UIColor { UIColor(named: "YP Red") ?? UIColor.red }
+    static var ypBlack: UIColor { UIColor(named: "YP Black") ?? UIColor.black }
+    static var ypBackground: UIColor { UIColor(named: "YP Background") ?? UIColor.darkGray }
+    static var ypGray: UIColor { UIColor(named: "YP Gray") ?? UIColor.gray }
+    static var ypWhite: UIColor { UIColor(named: "YP White") ?? UIColor.white}
 }
