+++ MovieQuiz/Models/AnswerButton.swift
@@ -1,0 +1,13 @@
+//
+//  AnswerButton.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+enum AnswerButton {
+    case yesButton
+    case noButton
+}
