+++ MovieQuiz/Models/ImagesCacheKeys.swift
@@ -1,0 +1,12 @@
+//
+//  ImagesCacheKeys.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+enum ImagesCacheKeys: String {
+    case cacheFileName = "iamgesCache"
+}
