+++ MovieQuiz/Models/LoadImagesError.swift
@@ -1,0 +1,12 @@
+//
+//  LoadImagesError.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+enum LoadImagesError {
+    case noCacheItem
+}
