+++ MovieQuiz/Models/LoadMoviesError.swift
@@ -1,0 +1,13 @@
+//
+//  LoadMoviesError.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+enum LoadMoviesError: Error {
+    case noCacheItem
+    case otherNetworkError(String)
+}
