+++ MovieQuiz/Models/MostPopularMoviesCacheKeys.swift
@@ -1,0 +1,13 @@
+//
+//  MostPopularMoviesCacheKeys.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+enum MostPopularMoviesCacheKeys: String {
+    case cacheFileName = "mainCache"
+    case moviesCacheKey = "structValue"
+}
