+++ MovieQuiz/Models/TestError.swift
@@ -1,0 +1,12 @@
+//
+//  TestError.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+enum TestError: Error { // тестовая ошибка
+    case test
+}
