+++ MovieQuiz/SceneDelegate.swift
@@ -1,7 +1,6 @@
 import UIKit
 
 final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
     var window: UIWindow?
 
     func scene(
