+++ MovieQuiz/Services/NetworkClient.swift
@@ -7,12 +7,8 @@
 
 import Foundation
 
-protocol NetworkRouting {
-    func fetch(url: URL, handler: @escaping (Result<Data, Error>) -> Void)
-}
-
 /// Отвечает за загрузку данных по URL
-struct NetworkClient: NetworkRouting {
+struct NetworkClient: NetworkRoutingProtocol {
     private enum NetworkError: Error {
         case codeError
     }
