+++ MovieQuiz/Services/NetworkRoutingProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  NetworkRoutingProtocol.swift
+//  MovieQuiz
+//
+//  Created by Сергей Кухарев on 24.04.2024.
+//
+
+import Foundation
+
+protocol NetworkRoutingProtocol {
+    func fetch(url: URL, handler: @escaping (Result<Data, Error>) -> Void)
+}
