+++ Tests/MovieQuizTests/ArrayTests.swift
@@ -10,7 +10,7 @@ import XCTest
 
 @testable import MovieQuiz
 
-class ArrayTests: XCTestCase {
+final class ArrayTests: XCTestCase {
     func testGetValueInRange() throws { // тест на успешное взятие элемента по индексу
         // Given
         let array = [1, 1, 2, 3, 5]
@@ -20,6 +20,7 @@ class ArrayTests: XCTestCase {
         XCTAssertNotNil(value)
         XCTAssertEqual(value, 2)
     }
+
     func testGetValueOutOfRange() throws { // тест на взятие элемента по неправильному индексу
         // Given
         let array = [1, 1, 2, 3, 5]
